Moved AssetId type definition
