+++ packages/config/src/balance/balance.ts
@@ -3,7 +3,7 @@ import '@moonbeam-network/api-augment';
 import { u128 } from '@polkadot/types';
 import { PalletBalancesAccountData } from '@polkadot/types/lookup';
 import { AssetSymbol } from '../constants';
-import { AssetId } from '../types';
+import { AssetId } from '../interfaces';
 import { BalanceFunction, BalancePallet } from './balance.constants';
 import {
   AssetsBalanceConfig,
