--- packages/config/src/types.ts
@@ -1,1 +1,0 @@
-export type AssetId = number | bigint | string;
